relocate Citations to manubot.cite.citations
